component/requirements.txt
awsiotsdk>=1.21.0
pyserial>=3.5
